# Optimize Prefixes and Merges