--- a/crates/milli/src/update/new/extract/mod.rs
+++ b/crates/milli/src/update/new/extract/mod.rs
@@ -7,8 +7,7 @@ mod vectors;
 
 use bumpalo::Bump;
 pub use cache::{
-    merge_caches, merge_caches_alt, transpose_and_freeze_caches, BalancedCaches,
-    DelAddRoaringBitmap,
+    merge_caches_sorted, transpose_and_freeze_caches, BalancedCaches, DelAddRoaringBitmap,
 };
 pub use documents::*;
 pub use faceted::*;