@Kerollmops 2024-12-04
/bench workloads/hackernews-add-new-documents.json workloads/hackernews-modify-*