@Kerollmops 2024-12-04
/bench workloads/movies.json workloads/hackernews.json