@Kerollmops 2024-12-04
bors abort

I want to maybe change the name of the `merge_alt` function and perform more benchmarks on larger machines.